Split functions into different files for readability
